Frontend UI changes
